Code for release 6.70
One primary change is to add -e option to replace comments with empty
comments, either /* */ or // (with nothing after it except the newline).
This makes some forms of analysis easier (and also means there isn't
trailing blanks in the output unless there were trailing blanks in the
input.

The other primary change affects the output from -c (print comments
instead of code); newlines are preserved in appropriate locations, so
that comments on separate lines are printed on separate lines.  This makes
the output easier to read.

One minor change is that the 'feature' names were changed and made more
systematic.